Move pkg/tools/histogram to knowledge/math
