+++ internal/knowledge/kpis/plugins/storage/storage_pool_cpu.go
@@ -11,8 +11,8 @@ import (
 	"github.com/cobaltcore-dev/cortex/internal/knowledge/db"
 	"github.com/cobaltcore-dev/cortex/internal/knowledge/extractor/plugins/storage"
 	"github.com/cobaltcore-dev/cortex/internal/knowledge/kpis/plugins"
+	"github.com/cobaltcore-dev/cortex/internal/knowledge/math"
 	"github.com/cobaltcore-dev/cortex/pkg/conf"
-	"github.com/cobaltcore-dev/cortex/pkg/tools"
 	"github.com/prometheus/client_golang/prometheus"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -74,7 +74,7 @@ func (k *NetAppStoragePoolCPUUsageKPI) Collect(ch chan<- prometheus.Metric) {
 	valueFunc := func(usage storage.StoragePoolCPUUsage) float64 {
 		return usage.MaxCPUUsagePct
 	}
-	hists, counts, sums := tools.Histogram(usages, buckets, keysFunc, valueFunc)
+	hists, counts, sums := math.Histogram(usages, buckets, keysFunc, valueFunc)
 	for key, hist := range hists {
 		ch <- prometheus.MustNewConstHistogram(k.storagePoolCPUUsageMax, counts[key], sums[key], hist)
 	}
@@ -84,7 +84,7 @@ func (k *NetAppStoragePoolCPUUsageKPI) Collect(ch chan<- prometheus.Metric) {
 	valueFunc = func(usage storage.StoragePoolCPUUsage) float64 {
 		return float64(usage.AvgCPUUsagePct)
 	}
-	hists, counts, sums = tools.Histogram(usages, buckets, keysFunc, valueFunc)
+	hists, counts, sums = math.Histogram(usages, buckets, keysFunc, valueFunc)
 	for key, hist := range hists {
 		ch <- prometheus.MustNewConstHistogram(k.storagePoolCPUUsageAvg, counts[key], sums[key], hist)
 	}
