+++ internal/knowledge/math/histogram_test.go
@@ -1,7 +1,7 @@
 // Copyright SAP SE
 // SPDX-License-Identifier: Apache-2.0
 
-package tools
+package math
 
 import (
 	"reflect"
